Added documentation, fixing diagrams, fixing tests and comments.
