+++ documentation/src/main/DocumentationTemplate.md
@@ -12,7 +12,8 @@
 #include "wiki/Splitter.md"
 #include "wiki/Knot.md"
 #include "wiki/ActionKnot.md"
-#include "wiki/ViewKnot.md"
+#include "wiki/ServiceKnot.md"
+#include "wiki/HandlebarsKnot.md"
 #include "wiki/Adapter.md"
 #include "wiki/ServiceAdapter.md"
 #include "wiki/ActionAdapter.md"
