+++ documentation/src/main/wiki/ServiceAdapter.md
@@ -1,5 +1,5 @@
 # Service Adapter
-Service Adapter is Component of a system, that mediate communication between Knot.x [[View Knot|ViewKnot]]
+Service Adapter is Component of a system, that mediate communication between Knot.x [[Service Knot|ServiceKnot]]
 and external services that deliver data injected into template. In short, Service Adapter acts as a
 `proxy` between Knot.x and external services.
 
@@ -12,7 +12,7 @@ Service Adapter accepts message with the following data:
 Result generated by Action Adapter must be a JsonObject with the fields as below:
 - `clientResponse` - Json Object, `body` field of this response is suppose to carry on the actual response from the mocked service,
 
-Find out more about contract described above in [[View Knot|ViewKnot]] section.
+Find out more about contract described above in [[Service Knot|ServiceKnot]] section.
 
 ## How to configure?
 Configuration of Service Adapter is specific to its implementation. You may see example configuration 
@@ -28,5 +28,5 @@ See [[Http Service Adapter|HttpServiceAdapter]] for more information.
 Please note, that this implementation is very generic and we recommend to create project-specific 
 adapters for any custom solution.
 
-Writing custom Service Adapter requires fulfilling [[View Knot|ViewKnot]] contract.
+Writing custom Service Adapter requires fulfilling [[Service Knot|ServiceKnot]] contract.
 Please refer also to [[Adapter|Adapter#how-to-extend]].
