+++ documentation/src/main/wiki/_Sidebar.md
@@ -13,7 +13,8 @@
   * [[Splitter|Splitter]]
   * [[Knot|Knot]]
     * [[Action Knot|ActionKnot]]
-    * [[View Knot|ViewKnot]]
+    * [[Service Knot|ServiceKnot]]
+    * [[Handlebars Knot|HandlebarsKnot]]
   * [[Adapter|Adapter]]
     * [[Service Adapter|ServiceAdapter]]
     * [[Action Adapter|ActionAdapter]]
