+++ knotx-knots/knotx-knot-action/src/main/resources/logback.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Knot.x - Reactive microservice assembler - View Knot
+    Knot.x - Reactive microservice assembler - Action Knot
 
     Copyright (C) 2016 Cognifide Limited
