+++ knotx-knots/knotx-knot-handlebars/src/main/java/com/cognifide/knotx/knot/templating/HandlebarsKnotConfiguration.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.cognifide.knotx.templating;
+package com.cognifide.knotx.knot.templating;
 
 import com.cognifide.knotx.knot.api.KnotConfiguration;
