+++ knotx-knots/knotx-knot-handlebars/src/main/java/com/cognifide/knotx/knot/templating/wrappers/TemplateHtmlFragment.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.cognifide.knotx.templating.wrappers;
+package com.cognifide.knotx.knot.templating.wrappers;
 
 import com.cognifide.knotx.fragments.Fragment;
 import com.cognifide.knotx.handlebars.JsonObjectValueResolver;
